# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NectarPin API",
    "description": "NectarPin 博客系统 API 接口文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "status": {
            "type": "integer",
            "description": "用户状态"
          },
          "role": {
            "type": "integer",
            "description": "用户角色"
          }
        },
        "required": ["id", "username", "email", "status", "role"]
      },
      "UpdateProfileRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "nickname": {
            "type": "string",
            "description": "昵称",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL",
            "maxLength": 500
          }
        }
      },
      "UserSessionItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "会话记录 ID" },
          "created_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "is_current": { "type": "boolean", "description": "是否为请求头 X-Refresh-Token 对应的会话" }
        },
        "required": ["id", "created_at", "expires_at", "is_current"]
      },
      "UserSessionListData": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserSessionItem" }
          }
        },
        "required": ["items"]
      },
      "RevokeOtherSessionsRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "当前设备持有的 refresh_token，用于保留本会话并撤销其余会话"
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["old_password", "new_password"],
        "properties": {
          "old_password": {
            "type": "string",
            "description": "当前密码（MD5 加密后的 32 位十六进制，与登录一致）",
            "minLength": 32,
            "maxLength": 32
          },
          "new_password": {
            "type": "string",
            "description": "新密码（MD5 加密后的 32 位十六进制）",
            "minLength": 32,
            "maxLength": 32
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "当前设备持有的 refresh_token"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "访问令牌"
          },
          "refresh_token": {
            "type": "string",
            "description": "刷新令牌（刷新成功后为新令牌，旧令牌作废）"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "access_token 过期时间（UTC）"
          },
          "token_type": {
            "type": "string",
            "description": "固定为 Bearer",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "过期时间（秒），兼容字段；服务端以 expires_at 为准"
          }
        },
        "required": ["access_token", "refresh_token", "expires_at"]
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "error": {
            "type": "string",
            "description": "详细错误信息"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/public/user/v1/token/refresh": {
      "post": {
        "tags": ["用户模块/公共接口/v1"],
        "summary": "刷新访问令牌",
        "description": "使用有效的 refresh_token 换取新的 access_token 与 refresh_token（旋转刷新令牌）",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "刷新成功" },
                    "data": { "$ref": "#/components/schemas/TokenResponse" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "刷新令牌无效或已过期",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "code": 401, "message": "刷新令牌无效或已过期" }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/protected/user/v1/profile": {
      "get": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "获取当前用户资料",
        "description": "获取当前登录用户的详细资料信息",
        "operationId": "getProfile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "获取成功" },
                    "data": { "$ref": "#/components/schemas/UserResponse" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "code": 401, "message": "未授权" } } }
          },
          "404": {
            "description": "用户不存在",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "post": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "更新当前用户资料",
        "description": "更新当前登录用户的昵称、邮箱与头像地址",
        "operationId": "updateProfile",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "更新成功" },
                    "data": { "$ref": "#/components/schemas/UserResponse" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "未授权",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "用户不存在",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "邮箱已被其他用户使用",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "code": 409, "message": "邮箱已被使用" } } }
          }
        }
      }
    },
    "/api/protected/user/v1/password": {
      "post": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "修改登录密码",
        "description": "校验当前密码后更新为新密码；传输格式与登录相同，为 MD5(明文) 的 32 位 hex",
        "operationId": "changePassword",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码已更新",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "密码已更新" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误、当前密码错误或与旧密码相同",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "未授权",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "用户不存在",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "服务器错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/protected/user/v1/logout/refresh-token": {
      "post": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "通过 Refresh Token 退出登录",
        "description": "使用 Refresh Token 退出登录",
        "operationId": "logoutByRefreshToken",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string", "description": "刷新令牌" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登出成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 200, "message": "登出成功" } } }
          },
          "400": {
            "description": "请求参数错误",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/protected/user/v1/sessions": {
      "get": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "列出登录会话",
        "description": "列出当前用户全部 refresh 会话。可选请求头 X-Refresh-Token 传入当前设备的 refresh 明文，用于标记 is_current",
        "operationId": "listUserSessions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "当前浏览器持有的 refresh_token 明文"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/UserSessionListData" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "查询失败",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/protected/user/v1/sessions/all": {
      "post": {
        "tags": ["用户模块/私有接口/v1"],
        "summary": "撤销所有其他设备的会话",
        "description": "请求体传入当前设备的 refresh_token；服务端校验后删除该用户其余 refresh 会话，仅保留当前这一条",
        "operationId": "revokeAllSessions",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RevokeOtherSessionsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "撤销成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 200, "message": "已撤销所有其他设备的登录会话" } } }
          },
          "400": {
            "description": "缺少或无效的 refresh_token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "未授权",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "撤销失败",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}